Pneumonia_Classification_CNN/pneumonia_app/requirements.txt
absl-py==2.3.1
astunparse==1.6.3
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
Flask==3.1.2
flatbuffers==25.2.10
gast==0.6.0
google-pasta==0.2.0
grpcio==1.74.0
h5py==3.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
keras==3.11.3
libclang==18.1.1
Markdown==3.8.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.3
namex==0.1.0
numpy==2.0.2
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pillow==11.3.0
protobuf==5.29.5
Pygments==2.19.2
requests==2.32.5
rich==14.1.0
six==1.17.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==1.17.3
